+++ PIR/src/passes/instCombine.cpp
@@ -870,6 +870,8 @@ namespace pcit::pir::passes{
 			case Expr::Kind::ASM:               return false;
 			case Expr::Kind::EXTRACT_ASM_VALUE: return false;
 			case Expr::Kind::ASM_VOID:          return false;
+			case Expr::Kind::WASM_MEMORY_GROW:  return false;
+			case Expr::Kind::WASM_MEMORY_SIZE:  return false;
 			case Expr::Kind::META_LOCAL_VAR:    return false;
 			case Expr::Kind::META_PARAM:        return false;
 		}
@@ -1278,6 +1280,9 @@ namespace pcit::pir::passes{
 			case Expr::Kind::ASM:               return false;
 			case Expr::Kind::EXTRACT_ASM_VALUE: return false;
 			case Expr::Kind::ASM_VOID:          return false;
+
+			case Expr::Kind::WASM_MEMORY_GROW: return false;
+			case Expr::Kind::WASM_MEMORY_SIZE: return false;
 
 			case Expr::Kind::META_LOCAL_VAR: return false;
 			case Expr::Kind::META_PARAM:     return false;